+++ src/app/local-storage/local-storage-service.module.ts
@@ -1,0 +1,11 @@
+import { NgModule } from '@angular/core';
+import { LocalStorageService } from './local-storage.service';
+
+@NgModule({
+    providers: [
+        LocalStorageService
+    ]
+})
+export class LocalStorageServiceModule {
+
+}