+++ src/app/projects-service/projects-service.module.ts
@@ -1,7 +1,12 @@
 import { NgModule } from '@angular/core';
 import { ProjectsService } from './projects.service';
+import { LocalStorageServiceModule } from '../local-storage/local-storage-service.module';
 
 @NgModule({
+    imports: [
+        // custom
+        LocalStorageServiceModule
+    ],
     providers: [
         ProjectsService
     ]