+++ src/app/projects-service/squishy-projects.ts
@@ -1,0 +1,3 @@
+import { SquishyProject } from "./squishy-project";
+
+export type SquishyProjects = { [key: string]: SquishyProject }