+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
  - `I18n` for english
+ - projects now stored in `localStorage`
  - `ExecuteDialog` for execution status, progress and cancel
 
 ## 0.1.0